+++ src/new/qurt/fcntl.rs
@@ -21,6 +21,9 @@ pub const O_NONBLOCK: c_int = 0x0800;
 pub const O_SYNC: c_int = 0x1000;
 pub const O_FSYNC: c_int = O_SYNC;
 pub const O_DSYNC: c_int = 0x1000;
+pub const O_DIRECTORY: c_int = 0x10000;
+pub const O_NOFOLLOW: c_int = 0x20000;
+pub const O_CLOEXEC: c_int = 0x80000;
 
 // fcntl() commands
 pub const F_DUPFD: c_int = 0;
@@ -31,6 +34,7 @@ pub const F_SETFL: c_int = 4;
 pub const F_GETLK: c_int = 5;
 pub const F_SETLK: c_int = 6;
 pub const F_SETLKW: c_int = 7;
+pub const F_DUPFD_CLOEXEC: c_int = 1030;
 
 // File descriptor flags
 pub const FD_CLOEXEC: c_int = 1;
@@ -43,7 +47,6 @@ pub const F_UNLCK: c_int = 2;
 // Functions
 extern "C" {
     pub fn open(pathname: *const c_char, flags: c_int, ...) -> c_int;
-    pub fn openat(dirfd: c_int, pathname: *const c_char, flags: c_int, ...) -> c_int;
     pub fn creat(pathname: *const c_char, mode: mode_t) -> c_int;
     pub fn fcntl(fd: c_int, cmd: c_int, ...) -> c_int;
 }
