+++ src/new/qurt/sys/mod.rs
@@ -1,4 +1,6 @@
 //! System headers (`sys/*`)
 
+pub(crate) mod mman;
+pub(crate) mod sched;
 pub(crate) mod stat;
 pub(crate) mod types;
